Relocate DataModels
