+++ config.yml
@@ -5,14 +5,14 @@ devices:
       host: 192.168.1.70 # Default IP for no DHCP
       contact_closure:
         port1: 
-          - name: "Device One" # What will be displayed on UI for control panel
-            pwr_on_state : 0   # After power loss what should be the port state (default: 0)
+          name: "Device One" # What will be displayed on UI for control panel
+          default_state : 0   # After power loss what should be the port state (default: 0)
         port2: 
-          - name: "Device Two"
-            pwr_on_state: 0
+          name: "Device Two"
+          default_state: 0
         port3: 
-          - name: "Device Three"
-            pwr_on_state : 0
+          name: "Device Three"
+          default_state : 0
 
 settings:
   web_api_server_port: 8000
