+++ itachweb/datamodels/ip2cc/__init__.py
@@ -1,0 +1,1 @@
+from .ip2ccDTO import *
