+++ itachweb/itach/ip2cc.py
@@ -1,66 +1,10 @@
 import re
 from typing import List
-from pydantic import BaseModel, ValidationInfo, field_validator
-from dataclasses import dataclass, replace
+
+from ..datamodels.ip2cc import IP2CCState
 from ..logger import syslog
 from .itach import ItachClient
 from ..error import check_response
-
-
-@dataclass
-class IP2CCState(BaseModel):
-    module: int
-    port: int
-    state: int
-
-
-@dataclass
-class IP2CCPortStates(BaseModel):
-    port1: int
-    port2: int
-    port3: int
-
-
-@dataclass
-class IP2CCPortUpdate(BaseModel):
-    device_id: int
-    module: int = 1
-    port: int
-    state: int
-
-    # @field_validator('device_id', 'module', 'port', 'state')
-    # @classmethod
-    # def validate_atts(cls, v: int, info: ValidationInfo):
-    #    if info.field_name == 'device_id':
-    #        if v < 0: raise ValueError(f'{v} is not a valid device id.')
-    #    elif info.field_name == 'module':
-    #        if v != 1: raise ValueError(f'{v} is not a valid module.')
-    #    elif info.field_name == 'port':
-    #         if not v in range(1,4): raise ValueError(f'{v} is not a valid port number.')
-    #    elif info.field_name == 'state':
-    #        if not v in range(2): raise ValueError(f'{v} is not a valid state.')
-    #    return v
-
-
-@dataclass
-class IP2CCPortDetail(BaseModel):
-    name: str
-    state: int
-
-
-@dataclass
-class IP2CCClosures(BaseModel):
-    port1: IP2CCPortDetail
-    port2: IP2CCPortDetail
-    port3: IP2CCPortDetail
-
-
-@dataclass
-class IP2CCDataModel(BaseModel):
-    id: int
-    name: str
-    host: str
-    contact_closure: IP2CCClosures
 
 
 class IP2CC(ItachClient):
@@ -113,7 +57,7 @@ class IP2CC(ItachClient):
         states = {}
         for i in range(1, 4):
             port = f"port{i}"
-            states[port] = int(self.get_state(1, i)["state"])
+            states[port] = int(self.get_state(1, i).state)
         return states
 
     def get_state(self, module, port):
@@ -233,7 +177,6 @@ class IP2CC(ItachClient):
         else:
             # (get|set)state,<module>:<port>,<state>
             mod_port = resp[1].split(":")
-            s = {"module": mod_port[0], "port": mod_port[1], "state": resp[2]}
-            return s
+            return IP2CCState(module=mod_port[0], port=mod_port[1], state=resp[2])
 
         return r
